requirements.txt
pytest
Django
djangorestframework
pytest-django